AFSB CONFESSION/requirements.txt
Flask
Pillow
instagrapi
python-dotenv
openai
